refactor: route internal imports through domain package facades
Consumers of domain.model and domain.plan now import from the package
namespace instead of reaching into concrete submodules (.constraints,
.table, .table_aspect, .actions, .diff). This applies the facade
convention already used by adapters/databricks/sql uniformly: external
callers import from the package; sibling modules inside a package still
import each other directly to avoid init-time cycles.

The domain.plan facade was previously dead - every caller dug into
.actions/.diff; it is now the single import path. Behaviour-preserving.
